Halfway through adding comments
