+++ Services/NeuralNetwork.Service/ActivationFunctions/ReLu.cs
@@ -2,8 +2,16 @@
 
 namespace NeuralNetwork.Service.ActivationFunctions;
 
+/// <summary>
+/// The ReLu activation function.
+/// </summary>
 public class ReLu : IActivationFunction
 {
+    /// <summary>
+    /// Apply the ReLu function to a signal input.
+    /// </summary>
+    /// <param name="Z">The sum of the signals entering the neuron and the bias.</param>
+    /// <returns>The signal input transformed by the ReLu function and outputted by the neuron (h).</returns>
     public double Calculate_h(double Z)
     {
         if (Z > 0)
@@ -13,6 +21,11 @@ public class ReLu : IActivationFunction
         return 0;
     }
 
+    /// <summary>
+    /// Calculate dh_dZ at the neuron for a ReLu function, used for back propagation.
+    /// </summary>
+    /// <param name="h">The output signal leaving the neuron during the most recent forward propagation.</param>
+    /// <returns>The gradient dh_dZ (used for back propagation).</returns>
     public double Calculate_dh_dZ(double h)
     {
         if (h > 0)
