+++ Services/NeuralNetwork.Service/Notes/Tasks.txt
@@ -1,0 +1,7 @@
+﻿Tasks with the NeuralNetwork.Service:
+
+1. The biases of neurons are not considered in forward or back propagation at the moment - they should be.
+
+2. Randomise the initial weight value of synapses.
+
+3. Implement an IWeightInitialiser interface and WeightInitialiser classes to use various weight initialisation methods.
